it4lia-ai-factory-evidence/requirements.txt
torch>=2.3
transformers>=4.40
datasets>=2.19
PyYAML>=6.0
numpy>=1.26
scikit-learn>=1.5
onnx>=1.16
onnxruntime>=1.18
fastapi>=0.110
uvicorn[standard]>=0.29
psutil>=5.9
pytest>=7.0
